Frontend: Don't load round info in root component
