+++ vue-app/src/views/ProjectList.vue
@@ -92,7 +92,7 @@ import Component from 'vue-class-component'
 import { FixedNumber } from 'ethers'
 import { DateTime } from 'luxon'
 
-import { RoundInfo } from '@/api/round'
+import { RoundInfo, getCurrentRound } from '@/api/round'
 import { Project, getProjects } from '@/api/projects'
 
 import ProjectListItem from '@/components/ProjectListItem.vue'
@@ -162,20 +162,23 @@ export default class ProjectList extends Vue {
     return this.$store.state.currentRound
   }
 
-  created() {
-    const roundAddress = this.$route.params.address || null
+  async created() {
+    const roundAddress = this.$route.params.address || this.$store.state.currentRoundAddress || await getCurrentRound()
     if (roundAddress && roundAddress !== this.$store.state.currentRoundAddress) {
       // Change current round and reload round info
-      this.$store.dispatch(UNWATCH_CART)
-      this.$store.dispatch(UNWATCH_CONTRIBUTOR_DATA)
-      this.$store.commit(SET_CONTRIBUTION, null)
-      this.$store.commit(SET_CONTRIBUTOR, null)
-      this.$store.commit(CLEAR_CART)
+      if (this.$store.state.currentRoundAddress) {
+        // Clear current round
+        this.$store.dispatch(UNWATCH_CART)
+        this.$store.dispatch(UNWATCH_CONTRIBUTOR_DATA)
+        this.$store.commit(SET_CONTRIBUTION, null)
+        this.$store.commit(SET_CONTRIBUTOR, null)
+        this.$store.commit(CLEAR_CART)
+      }
       this.$store.commit(SET_CURRENT_ROUND_ADDRESS, roundAddress)
       ;(async () => {
         await this.$store.dispatch(LOAD_ROUND_INFO)
         if (this.$store.state.currentUser) {
-          // Reload user data when switching between rounds
+          // Load user data if already logged in
           this.$store.dispatch(LOAD_USER_INFO)
           this.$store.dispatch(LOAD_CART)
           this.$store.dispatch(LOAD_CONTRIBUTOR_DATA)
